==> R/p00_characterise_Dejong_data_04_04_2017.Rmd <==
---
title: "CellMix"
author: "DJL"
date: "17/02/2016"
output:
  word_document:
    fig_height: 6
---

#CellMix approach:

##Install and load Libraries:
```{r Libraries, message=FALSE,warning=FALSE}
rm(list = ls())
dev.off()
library(lumi)
library(reshape)
library(plyr)
library(dplyr)
library(ggplot2)
library(tableone)
library(stargazer)
#library(illuminaHumanv4.db)
library(illuminaHumanv3.db)



```


##Set directories:
```{r directory}

data_dir <-"./data/"
P0_output_dir <-"./P0_Characterise/output/"
P00_output_dir <-"./P00_Characterise_Dejong/output/"
P00_figs_dir <-"./P00_Characterise_Dejong/figs/"

```



##Load gene expression data (Should be LumiBatch object):
```{r}

lumidata_full<-"Schizophrenia_analysis_files.Rdata"
load(paste(data_dir,lumidata_full,sep=""))


lumidata<-"GAP_FEP_Full_Gene_Expression_Data_Linear.RData"
load(paste(data_dir,lumidata,sep=""))

lumidata_full<-"GAP_FEP_eset_linear_full_adj_Data.RData"
load(paste(P0_output_dir,lumidata_full,sep=""))


```

##Functions:
```{r functions}

## Table function 1 demographics
table_dem_fun<-function(pdata,table_name,stratify,listVars,catVars){
  table1 <- CreateTableOne(vars = listVars, data = pdata, factorVars = catVars,strata=c(stratify),includeNA = T)
  table1print<-print(table1)
  table1print<-table1print[,-length(names(data.frame(table1print)))]
  write.csv(table1print, file=paste(P0_output_dir,"no_latex_",table_name,sep=""),row.names = TRUE, col.names = TRUE,quote=FALSE,sep = ",")
  write.csv(stargazer(table1print,summary=FALSE), file=paste(P0_output_dir,"latex_",table_name,sep=""),row.names=F,col.names = F,sep="")
}

```

##Gene symbol translation:
```{r }
expression_chip="illuminaHumanv3"


#create NUID translation for Dejong data
mapped_probes <- mappedkeys(eval(parse(text = paste(expression_chip, "NUID", sep=""))))
probe_nuID_mapping <- as.data.frame(eval(parse(text = paste(expression_chip, "NUID", sep="")))[mapped_probes])

```


#Clean dataframes:
```{r Clean Dataframes dejong}

#Get GAP translation Dataframes
GAP_ALL<-eset_bg_log2_rsn_SVA@featureData@data
GAP_final_probes_linear<-eset_linear_full_adj@featureData@data

#Add Illumina v 3 probe ids
colnames(probe_nuID_mapping)<-c("IlluminaV3","nuID")
GAP_full_DejongFDF<-inner_join(GAP_ALL, probe_nuID_mapping, by = "nuID")
GAP_final_DejongFDF<-inner_join(GAP_final_probes_linear, probe_nuID_mapping, by = "nuID")



#Expressiontable Full De jong probes
Gx_Columns_dejong<-colnames(Schizophrenia_full_expression_table)

GAP_full_DejongFDF<-GAP_full_DejongFDF[GAP_full_DejongFDF$IlluminaV3%in%Gx_Columns_dejong,] #smaller again


#dejong_nuIDs mapped
dejong_nuID<-probe_nuID_mapping[probe_nuID_mapping$IlluminaV3%in%Gx_Columns_dejong,]

#probe translate subset to dejong probes
GAP_final_dejongprobes<-GAP_final_DejongFDF[GAP_final_DejongFDF$nuID%in%dejong_nuID$nuID,]




#subset to gene expression dataframe
DejongGxsmall<-Schizophrenia_full_expression_table[,c("Diagnosis",GAP_final_dejongprobes$IlluminaV3)]
colnames(DejongGxsmall)<-c("Diagnosis",GAP_final_dejongprobes$TargetID)#Change colnames to Gene Symbol

#Find probes that are still missing to see if they can be identified. 
Dejong_Missing<-GAP_final_probes_linear[!GAP_final_probes_linear$TargetID%in%colnames(DejongGxsmall),]

Dejong_Missing$TargetID

#Use Entrenz Gene Id to find missing probes
colID<-"ENTREZ_GENE_ID"
table(GAP_full_DejongFDF[,colID]%in%Dejong_Missing[,colID])
dejong_101_added<-GAP_full_DejongFDF[GAP_full_DejongFDF[,colID]%in%Dejong_Missing[,colID],] 
dejong_101_added<-dejong_101_added[complete.cases(dejong_101_added$ENTREZ_GENE_ID),] #remove NAs

#create EntrezIdslist
DejongGx_nuIDs<-Gx_Columns_dejong[Gx_Columns_dejong%in%GAP_final_dejongprobes$IlluminaV3]
DejongGx_entrezIDs<-Gx_Columns_dejong[Gx_Columns_dejong%in%dejong_101_added$IlluminaV3]



#back into full translate DF
New_translate<-GAP_full_DejongFDF[GAP_full_DejongFDF$IlluminaV3%in%c(DejongGx_nuIDs,DejongGx_entrezIDs),]
New_translate<-New_translate[!duplicated(New_translate$TargetID),]

#Missing Probes Dejong use inner join. 
Missing_probes_Dejong<-GAP_final_probes_linear[!GAP_final_probes_linear$TargetID%in%New_translate$TargetID,]
dim(Missing_probes_Dejong)
  

#Feature Data final
Final_GAP_Fdata_IDs<-GAP_final_probes_linear[GAP_final_probes_linear$TargetID%in%New_translate$TargetID,] #Final Gap Probes
Final_Dejong_Fdata_IDs<-New_translate[New_translate$TargetID%in%GAP_final_probes_linear$TargetID,] #final dejong probes


all.equal(Final_Dejong_Fdata_IDs$TargetID,Final_GAP_Fdata_IDs$TargetID) # check if all probes are equal. 
dim(Final_GAP_Fdata_IDs) #4220
dim(Final_Dejong_Fdata_IDs) #4220

#Full Join for final DF
GAP_Dejong_Full<-full_join(GAP_ALL, probe_nuID_mapping, by = "nuID")
dim(GAP_Dejong_Full)

#Add Column with true false by nuIDs for both Gap and Dejong data
GAP_Dejong_Full$GAP_GAPDJ<-GAP_Dejong_Full$nuID%in%Final_GAP_Fdata_IDs$nuID
GAP_Dejong_Full$DeJong_GAPDJ<-GAP_Dejong_Full$nuID%in%Final_Dejong_Fdata_IDs$nuID


table(GAP_Dejong_Full$GAP_GAPDJ)
table(GAP_Dejong_Full$DeJong_GAPDJ)


#Final Fdata
GAP_Dejong_Full_Fdata<-GAP_Dejong_Full

#Make Dejong Gx dataframe

#subset to gene expression dataframe
Gx_dejong<-Schizophrenia_full_expression_table[,c("Diagnosis",GAP_Dejong_Full[GAP_Dejong_Full$DeJong_GAPDJ == "TRUE",]$IlluminaV3)]
colnames(Gx_dejong)<-c("Diagnosis",GAP_Dejong_Full[GAP_Dejong_Full$DeJong_GAPDJ == "TRUE",]$TargetID)#Change colnames to Gene Symbol



#Phenodata
Pheno_dejong<-Schizophrenia_phenotype_data[,c(1,8:10,36)]

```

#Write rdata:
```{r Clean Dataframes dejong}


save(Pheno_dejong,Gx_dejong,Missing_probes_Dejong,GAP_Dejong_Full_Fdata,file=paste(P00_output_dir,"Dejong_reduced_Gx_data_and_pheno.RData",sep=""))


```


#Write demographics:
```{r demographics}

str(Schizophrenia_phenotype_data)
names(Schizophrenia_phenotype_data)
head(Schizophrenia_phenotype_data)

mldata=Schizophrenia_phenotype_data[,c(32:33,36)]
names(mldata)=c("Age","Gender","Phenotype")

t_title="_Dejong_all"
latex="_latex"
filename_demo=paste(output_dir,project_id,project_name,t_title,"_Demographics.csv",sep="")
#paste(output_dir,project_id,"not_latex","_",t_title,"_","Table_CM_1_Demographics.csv",sep="")


listVars <- c("Gender","Age")
catVars <- c("Gender")
table1 <- CreateTableOne(vars = listVars, data = mldata, factorVars = catVars,strata=c("Phenotype"),includeNA = T)
table1print<-print(table1, quote = FALSE, noSpaces = TRUE, printToggle = FALSE)
#not latex
write.csv(table1print, file=filename_demo,row.names = TRUE, col.names = TRUE,quote=FALSE,sep=",")

filename_demo_latex=paste(output_dir,project_id,project_name,latex,"_",t_title,"_Demographics.csv",sep="")
#paste(output_dir,project_id,latex,"_",t_title,"_","Table_CM_1_Demographics.csv",sep="")
#stargazer latex tables
table1sg<-stargazer(table1print[,1:3],summary=FALSE, rownames=T)
write.csv(table1sg,filename_demo_latex,sep="",row.names=F,col.names = F)

```
