==> R/p0_CellMix_GAP_FEP_data_07_02_2017.Rmd <==
---
title: "CellMix"
author: "DJL"
date: "17/02/2016"
output:
  word_document:
    fig_height: 6
---

TODO: 
PRS - Adjust using Principle components. 
Get Residuals.
Center Scale. 
Remove outliers. 

#CellMix approach:

##Install and load Libraries:
```{r Libraries, message=FALSE,warning=FALSE}
rm(list = ls())
dev.off()

library(CellMix)
library(lumi)
library(reshape)
library(plyr)
library(dplyr)
library(ggplot2)


```


##Set directories:
```{r directory}

data_dir <-"./data/"
P0_output_dir <-"./P0_Characterise/output/"
P0_figs_dir <-"./P0_Characterise/figs/"

```

##Load gene expression data (Should be LumiBatch object):
```{r}

lumidata<-"GAP_FEP_Full_Gene_Expression_Data_Linear.RData"
load(paste(data_dir,lumidata,sep=""))
PRS_data<-read.csv(paste(data_dir,"GAP_FEP_Polygenic_Risk_scores.csv",sep=""))
PRS_eigenvalues_data<-read.csv(paste(data_dir,"GAP_eigenvectors_from_Vangelis_04_03_2015.csv",sep=""),na.strings=c("#N/A","notfound"))

```


##Subset to good probes:
```{r}

dim(eset_bg_log2_rsn_SVA)
eset_bg_log2_rsn_SVA_Good<-eset_bg_log2_rsn_SVA
#make exprs data with good probes
exprs_data<-exprs(eset_bg_log2_rsn_SVA)
feature_data<-fData(eset_bg_log2_rsn_SVA)
#get good probe
good_fdata<-filter(feature_data, good_probe =="TRUE")
#good_fdata_nHS<-good_fdata[!grepl("^HS\\.",good_fdata$TargetID),]
#good_fdata_nHSLoc<-good_fdata_nHS[!grepl("^LOC",good_fdata_nHS$TargetID),]

exprs_into_lumibatch<-exprs_data[rownames(exprs_data)%in%good_fdata$nuID,]

################ REMOVE DUPLICATES BY SELCTING HIGHEST AVERAGE EXPRESSED PROBE ###############################
#rowmeans
exprsmean<-rowMeans(exprs_into_lumibatch)
#add id
rownames(good_fdata)<-good_fdata$nuID
exprsmean_ids<-merge(as.data.frame(exprsmean),good_fdata[,c("nuID","TargetID")],by="row.names")
# order data frame by truncated probe id and then expression level
exprsmean_ids<-exprsmean_ids[order(exprsmean_ids$TargetID, -exprsmean_ids$exprsmean), ]

# remove all duplicate probe id - keep one with highest mean expression
exprsmean_ids_unique<-exprsmean_ids[!duplicated(exprsmean_ids$TargetID),]

#reduce exprs set again
exprs_into_lumibatch2<-exprs_into_lumibatch[exprsmean_ids_unique$Row.names,]

good_fdata_small<-good_fdata[exprsmean_ids_unique$Row.names,]
#check they are same order
all.equal(good_fdata_small$nuID,rownames(exprs_into_lumibatch2))



#LUMIBATCH made
eset_bg_log2_rsn_SVA_Good<-eset_bg_log2_rsn_SVA[rownames(exprs_into_lumibatch2),]
fData(eset_bg_log2_rsn_SVA_Good)<-good_fdata_small

save(eset_bg_log2_rsn_SVA_Good,file=paste(P0_output_dir,"GAP_FEP_small_Gene_Expression_Data.RData",sep=""))

```



##GED BLOOD ABBAS WHOLE BLOOD:
gedBlood is a meta function. It uses a standard set of functions included in CellMix to generate results. It seems to be able to handle nuIDs. gedBlood automatically adjusts gene IDs so I suspect it searches the LumiBatch file for IDs it recognises and matches them with whatever dataset it uses. 
In this case the ABBAS blood atlas is used. 
CLsubset allows you to choose between WB for whole blood and PBMCs. 
verbose simply tells you the steps it takes which would usually have to be performed manually in the package. 
These settings just use a linear model based on gene expression values for different cell types from the ABBAS blood atlas (Abbas et al. 2009)* to estimate cell proportions in each sample.
This is after preprocessing, so we only use the approx. 5000 probes that passed QC.
Cellmix has a lot more functions and ways to estimate cell proportions, but after looking through all of it, you could probably make it into a study by itself. Some of the methods seem to be quite computationally intensive. See table 2 on page 26 [here](http://web.cbio.uct.ac.za/~renaud/CRAN/web/CellMix/vignettes/Introduction.pdf) for list of approaches. 

^*Abbas AR, Wolslegel K, Seshasayee D, Modrusan Z and Clark HF (2009). "Deconvolution of blood microarray data identifies cellular activation patterns in systemic lupus erythematosus."^

###GEDBlood Function:
```{r}
res_all <- gedBlood(eset_bg_log2_rsn_SVA_Good, CLsubset = "WB", verbose = TRUE)


#Extract cell proportions from res_all
wbloodprop<-coef(res_all)

#Remove rows with sum of 0. Rows represent cell types. If a cell type has a sum of 0 across all samples I exclude it. 

reduced_props<-wbloodprop[apply(wbloodprop, 1, function(x) !all(x==0)),]

```

"reduced_probs" contains all the releveant cell proportions and is written into a CSV file which I later merge with my phenotype data in my LIMMA script. 
They are then added as covariates. 


##Check for differences between cases and controls:
Here I split the blood proportion data in case and control so I can plot it using ggplot2. 

```{r}

#Get Case Control Status
pheno_data <- pData(eset_bg_log2_rsn_SVA_Good)

#transform data
reduced_props<-data.frame(t(reduced_props))
#check samples in correct order
all.equal(rownames(reduced_props),rownames(pheno_data))
#add pheno data
reduced_probs_pheno<-cbind(reduced_props,pheno_data)
names(reduced_probs_pheno)

#Melt Data for ggplot
mwbdata <- melt(reduced_probs_pheno, id=c(11:71))
head(mwbdata)



#Graph data
bloodgraph<-ggplot(mwbdata,aes(x=variable,y=value,fill=Gender), ) +  
  geom_boxplot()+
  ggtitle("Blood Cell Proportions")+
  theme(plot.title = element_text(lineheight=.8, face="bold"))+
  xlab("Cell Type")+ylab("Percentage for each individual")
bloodgraph


#select plots to make
plotnames<-names(reduced_probs_pheno[c("Phenotype","Gender","Tobacco","Ethnicity","Medication","tech.Sentrix.Barcode","tech.Date_Washing_and_scanning","tech.Date_Quantitation_by_RiboGreen","tech.SampleSection")])

names(reduced_probs_pheno)

#make function
f <- function(mwbdata, fill_name) {
    bloodgraph<-ggplot(mwbdata,aes_string(x="variable",y="value",fill=fill_name), ) +  
    geom_boxplot()+
    ggtitle("Blood Cell Proportions")+
    theme(plot.title = element_text(lineheight=.8, face="bold"))+
    xlab("Cell Type")+ylab("Percentage for each individual")
    print(bloodgraph)
}

#Create loop for all variables and save
CellMixgraph<-"Cell_Mix_Boxplots.pdf"
pdf(paste(P0_figs_dir,CellMixgraph,sep=""))
for (fill_number in 1:length(plotnames)){
  fill_name<-plotnames[fill_number]
  f(mwbdata,fill_name)  
  
}
dev.off()

  
```

##Statistics:
#Looks like Tc and neutro are significant. I will adjust for that. 
```{r, tidy=TRUE}

names(mwbdata)
stats_pheno<-ddply(mwbdata,"variable",
      function(x) {
          w <- wilcox.test(value~Phenotype,data=x)
          with(w,data.frame(statistic,p.value))
      })

stats_pheno




```

##Add PRS, TC and neutro to phenodata:
```{r, tidy=TRUE}

cor(reduced_probs_pheno[1:10])

## add PRS
reduced_probs_pheno_prs=merge(reduced_probs_pheno,PRS_data[,-c(2:3)],by="sampleID")
names(reduced_probs_pheno)
names(reduced_probs_pheno_prs[,c(2:11,1,12:81)])
#Get everything in the right order
reduced_probs_pheno_prs<-reduced_probs_pheno_prs[,c(2:11,1,12:81)]
#second order (this is because I messed up a bit.)
reduced_probs_pheno_prs_order<-reduced_probs_pheno_prs[,c(11:71,1,10,72:81)]

## add columns for Dataframe for later use.
## Things to Change: PRS col names, ICD10 and DSMIV (control + collapse (2 extra columns),also change colnames),   
names(reduced_probs_pheno_prs_order)
head(reduced_probs_pheno_prs_order)

## 1) change PRS colnames
Prscolnames<-c("PRS_5e08", "PRS_1e05", "PRS_1e04", "PRS_0.001", "PRS_0.01", "PRS_0.05", "PRS_0.1", "PRS_0.2", "PRS_0.5", "PRS_1")
colnames(reduced_probs_pheno_prs_order)[64:73]<-Prscolnames

## 2.1) Change DSM and ICD10 colnames
diagnosiscols<-c("dsmiv","icd10")
colnames(reduced_probs_pheno_prs_order)[18:19]<-diagnosiscols

## 2.2) add control to dsmiv and icd.

reduced_probs_pheno_prs_order$Phenotype[reduced_probs_pheno_prs_order$Phenotype == "control"] <- "Control"
reduced_probs_pheno_prs_order$dsmiv[reduced_probs_pheno_prs_order$Phenotype == "Control"] <- "Control"
reduced_probs_pheno_prs_order$icd10[reduced_probs_pheno_prs_order$Phenotype == "Control"] <- "Control"

reduced_pheno_final<-reduced_probs_pheno_prs_order %>% mutate(ICD_DSM=ifelse(dsmiv=="Schizophrenia" | icd10 == "Schizophrenia","Schizophrenia",
                                                                ifelse(dsmiv =="Control", "Control", "Other_Psychosis")))


#########################################################
#Adjust PRS for Principle components

#New matrix Remove mising values
names(reduced_pheno_final)
head(reduced_pheno_final)
PRS_to_adjust<-reduced_pheno_final[c(1,6,9,64:73)]
rownames(PRS_to_adjust)<-reduced_pheno_final$sampleID

#remove columns not needed
PRS_to_adjust<-PRS_to_adjust[complete.cases(PRS_to_adjust),]


##get principle components to use for adjustment
#Reduce
PRS_eigenvalue_adjust<-PRS_eigenvalues_data[c(2,3,9:18)]
PRS_eigenvalue_adjust<-PRS_eigenvalue_adjust[complete.cases(PRS_eigenvalue_adjust),]
#Reduce to relevant Samples
PRS_eigenvalue_adjust<-PRS_eigenvalue_adjust[PRS_eigenvalue_adjust$GAP_ID%in%PRS_to_adjust$gap_id,]
colnames(PRS_eigenvalue_adjust)[2]<-"gap_id"
#change to integer
PRS_eigenvalue_adjust$gap_id<-as.integer(as.character(PRS_eigenvalue_adjust$gap_id))
#combine PRS values and eigenvalues.
Combined_PRS<-inner_join(PRS_eigenvalue_adjust,PRS_to_adjust,by="gap_id")

#add rownames
rownames(Combined_PRS)<-Combined_PRS$sampleID

#Make adjustment data and pdata
names(Combined_PRS)
PRS_matrix<-t(Combined_PRS[,15:24])

#Model For matrix, have samples as columns. variables as rows.
#mod = model.matrix(~eigen_1+eigen_2,data=Combined_PRS)
formula_formod<-as.formula(paste("~", paste(paste("eigen_",1:10,sep=""), collapse="+")))
mod = model.matrix(formula_formod,data=Combined_PRS)
fit = lm.fit(mod,t(PRS_matrix))
#add residuals to average expression data
PRS_adjusted<-t(fit$residuals)+apply(PRS_matrix,1,mean)

#clean PRS_adjusted
PRS_adjusted<-t(PRS_adjusted)
colnames(PRS_adjusted)<-paste(colnames(PRS_adjusted),"_adj",sep="")
PRS_adjusted<-as.data.frame(PRS_adjusted)
PRS_adjusted$sampleID<-rownames(PRS_adjusted)

PRS_adjusted_full<-cbind(Combined_PRS,PRS_adjusted)

#combine with orginal data
reduced_pheno_final_adj_PRS<-full_join(reduced_pheno_final,PRS_adjusted,by="sampleID")



ggplot(data = reduced_pheno_final_adj_PRS, 
      aes(x = ICD_DSM, y=PRS_0.1_adj,colour = ICD_DSM)) +
      geom_jitter() +
      geom_boxplot()+
      facet_wrap(~ Ethnicity)

title = "PRS vs Psychosis categories"
ggplot(data = reduced_pheno_final_adj_PRS, 
      aes(x = ICD_DSM, y=PRS_1,colour = ICD_DSM)) +
      geom_jitter() +
      geom_boxplot()+
      facet_wrap(~ Ethnicity)



#########################################################


pData(eset_bg_log2_rsn_SVA_Good)<-reduced_pheno_final_adj_PRS
names(pData(eset_bg_log2_rsn_SVA_Good))
save(eset_bg_log2_rsn_SVA_Good,file=paste(P0_output_dir,"GAP_FEP_small_Gene_Expression_Data.RData",sep=""))
```




##Linear model to remove covariate effect: done 2 times. Once to remove just blood, once to also consider ethncity, age, sex.
```{r, tidy=TRUE}


#Input
pData_rAESB<-pData(eset_bg_log2_rsn_SVA_Good)
exprs_rAESB<-exprs(eset_bg_log2_rsn_SVA_Good)


######################### First Step: JUST Blood differences#########################
mod = model.matrix(~Tc+neutro,data=pData_rAESB)
fit = lm.fit(mod,t(exprs_rAESB))
#add residuals to average expression data
exprs_res_cor_adj<-t(fit$residuals)+apply(exprs_rAESB,1,mean)
exprs_res_cor_adj[1:10,1:10]

#make Lumibatch
exprs_into_lumibatch2<-exprs_res_cor_adj
#check they are same order
all.equal(rownames(fData(eset_bg_log2_rsn_SVA_Good)),rownames(exprs_into_lumibatch2))
#LUMIBATCH made
eset_linear_adj<-eset_bg_log2_rsn_SVA_Good
#Expression data put in
exprs(eset_linear_adj)<-exprs_into_lumibatch2

#save
save(eset_linear_adj,file=paste(P0_output_dir,"GAP_FEP_eset_linear_adj_Data.RData",sep=""))


######################### Second Step: remove ethnicity, age, gender + tc + neutro #########################

mod = model.matrix(~Tc+neutro+as.factor(Gender)+Age+as.factor(Ethnicity),data=pData_rAESB)
fit = lm.fit(mod,t(exprs_rAESB))
#add residuals to average expression data
exprs_res_cor_adj<-t(fit$residuals)+apply(exprs_rAESB,1,mean)
exprs_res_cor_adj[1:10,1:10]


#make Lumibatch
exprs_into_lumibatch_fadj<-exprs_res_cor_adj
#check they are same order
all.equal(rownames(fData(eset_bg_log2_rsn_SVA_Good)),rownames(exprs_into_lumibatch_fadj))
#LUMIBATCH made
eset_linear_full_adj<-eset_bg_log2_rsn_SVA_Good
#Expression data put in
exprs(eset_linear_full_adj)<-exprs_into_lumibatch_fadj

#save
save(eset_linear_full_adj,file=paste(P0_output_dir,"GAP_FEP_eset_linear_full_adj_Data.RData",sep=""))


# check for equallity
all.equal(exprs(eset_linear_full_adj),exprs(exprs_res_cor_adj)) ## should be true
all.equal(exprs(eset_linear_full_adj),exprs(eset_linear_adj)) ## False. Some difference due to age ethnicity etc. 

```
