==> R/p0_characterise_GAP_FEP_data_07_02_2017.Rmd <==
---
title: "CellMix"
author: "DJL"
date: "17/02/2016"
output:
  word_document:
    fig_height: 6
---

#CellMix approach:

##Install and load Libraries:
```{r Libraries, message=FALSE,warning=FALSE}
rm(list = ls())
dev.off()
library(lumi)
library(reshape)
library(plyr)
library(dplyr)
library(ggplot2)
library(tableone)
library(stargazer)

```


##Set directories:
```{r directory}

data_dir <-"./data/"
P0_output_dir <-"./P0_Characterise/output/"
P0_figs_dir <-"./P0_Characterise/figs/"

```

##Load gene expression data (Should be LumiBatch object):
```{r}

#lumidata<-"GAP_FEP_eset_linear_adj_Data.RData"
#load(paste(P0_output_dir,lumidata,sep=""))
lumidata_full<-"GAP_FEP_eset_linear_full_adj_Data.RData"
load(paste(P0_output_dir,lumidata_full,sep=""))

```

##Functions:
```{r}

## Table function 1 demographics
table_dem_fun<-function(pdata,table_name,stratify,listVars,catVars){
  table1 <- CreateTableOne(vars = listVars, data = pdata, factorVars = catVars,strata=c(stratify),includeNA = T)
  table1print<-print(table1)
  table1print<-table1print[,-length(names(data.frame(table1print)))]
  write.csv(table1print, file=paste(P0_output_dir,"no_latex_",table_name,sep=""),row.names = TRUE, col.names = TRUE,quote=FALSE,sep = ",")
  write.csv(stargazer(table1print,summary=FALSE), file=paste(P0_output_dir,"latex_",table_name,sep=""),row.names=F,col.names = F,sep="")
}

```



##Demographics:
```{r}
str(phenodat)

phenodat<-pData(eset_linear_full_adj)
phenodat[64:73]<-scale(phenodat[64:73])

phenodat
names(phenodat)

#######Table1: Demographics
##variables
lVars <- c("Gender","Age", "Ethnicity","BMI","Tobacco",names(phenodat)[65:73])
cVars <- c("Ethnicity","Tobacco")
lVars2 <- c("Medication","dsmiv","icd10","ICD_DSM","PanssScore","PanssPositive","PanssNegative","PanssPsycho")
cVars2 <- c("Medication","dsmiv","icd10","ICD_DSM")


##table statified by case control
stratVar <- "Phenotype" 
table_name <-"Table_1_Demographics.csv"
table_dem_fun(phenodat,table_name,stratVar,lVars,cVars)
table_name <-"Table_2_Clinicalinformation.csv"
table_dem_fun(phenodat,table_name,stratVar,lVars2,cVars2)

##table statified by Diagnosis consensus.
stratVar <- "ICD_DSM" 
table_name <-"Table_1_SczCat_Demographics.csv"
table_dem_fun(phenodat,table_name,stratVar,lVars,cVars)
table_name <-"Table_2_SczCat_Clinicalinformation.csv"
table_dem_fun(filter(phenodat,Phenotype=="FEP"),table_name,stratVar,lVars2,cVars2)

##table statified by Diagnosis consensus.
stratVar <- "ICD_DSM" 
table_name <-"Table_1_SczCat_white_Demographics.csv"
table_dem_fun(filter(phenodat,Ethnicity =="White"),table_name,stratVar,lVars,cVars)
table_name <-"Table_2_SczCat_white_Clinicalinformation.csv"
table_dem_fun(filter(phenodat,Phenotype=="FEP" & Ethnicity =="White"),table_name,stratVar,lVars2,cVars2)




```

#Visual Summary of data 
```{r data summary}
#gender 
title = "Gender in FEP and Controls"
ggplot(data = phenodat, aes(x = Phenotype, colour = Gender, fill = Gender)) +
    geom_bar(alpha = 0.1)+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))


#Age
title = "Age differences between cases and controls"
ggplot(data = phenodat, aes(x = Phenotype, y=Age, colour = Phenotype, fill = Phenotype)) +
      geom_boxplot(alpha = 0)+
      geom_jitter(alpha = 0.3)+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))


#Age broken by gender
title = "Gender and Age differences between cases and controls"
ggplot(data = phenodat, aes(x = Phenotype, y=Age, colour = Phenotype, fill = Phenotype)) +
      geom_boxplot(alpha = 0)+
      facet_wrap(~ Gender)+
      geom_jitter(alpha = 0.3)+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))



#ethnicty 
title = "Ethnicity breakdown"
ggplot(data = phenodat, aes(x = Phenotype, colour = Gender, fill =Gender)) +
    geom_bar(alpha = 0.1)+
    facet_wrap(~ Ethnicity)+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))



#bmi 
title = "BMI differences between cases and controls"
ggplot(data = phenodat, aes(x = Phenotype  , y = BMI, colour = Gender, fill =Gender)) +
  geom_boxplot(alpha = 0.1)+
  facet_wrap(~ Gender)+ 
  geom_jitter(alpha = 0.3)+
  ggtitle(title)

ggsave(paste(P0_figs_dir,title,".png",sep=""))


#bmi 
title = "Medication and Weight"
ggplot(data = filter(phenodat,Medication!="CONTROL"), 
      aes(x = Medication, y=BMI  , colour = Medication)) +
      geom_point() + 
      geom_boxplot()+
      coord_flip()+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))




#tobacco
title = "Smoking in cases and controls"
ggplot(data = phenodat, 
      aes(x = Tobacco, colour = Tobacco, fill =Tobacco))+
      geom_bar(alpha = 0.1)+
      facet_wrap(~ Phenotype)+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))

#medication 
title = "Medication for FEP patients"
ggplot(data = filter(phenodat,Medication!="CONTROL"), 
      aes(x = Medication , colour = Medication)) +
      geom_bar(alpha = 0.1)+
      coord_flip()+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))


names(phenodat)
#medication vs PRS
title = "Medication vs PRS"
ggplot(data = filter(phenodat,Medication!="CONTROL"), 
      aes(x = Medication, y=PRS_1  , colour = Medication)) +
      geom_point() + 
      geom_boxplot()+
      coord_flip()+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))

#medication vs PRS
title = "PRS vs Ethnicity"
ggplot(data = phenodat, 
      aes(x = Phenotype, y=PRS_1  , colour = Ethnicity)) +
      geom_jitter() + 
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))

#medication vs PRS
title = "Medication vs Ethnicity"
ggplot(data = filter(phenodat,Ethnicity=="White"), 
      aes(x = , y=PRS  , colour = Ethnicity)) +
      geom_jitter() +
      geom_boxplot()+
      ggtitle(title)

#medication vs PRS
title = "PRS vs Psychosis categories"
ggplot(data = phenodat, 
      aes(x = ICD_DSM, y=PRS_1  , colour = ICD_DSM)) +
      geom_jitter() +
      geom_boxplot()+
      facet_wrap(~ Ethnicity)+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))


#medication vs PRS
title = "PRS vs Psychosis categories"
ggplot(phenodat, aes(PRS_1,fill = ICD_DSM, colour = ICD_DSM)) +
  geom_density(alpha=0.01)+
  facet_wrap(~ Ethnicity)+
  ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))


```

#Visual Summary of data 
```{r data summary}
#melt by panss
phenodat_melt<-melt(phenodat, id = c("gap_id","SampleID","sampleID","Phenotype","Gender","Age","Ethnicity","BMI","Pol_0.05_GAP_all_strict_excl_WTCCC2","Tobacco","Medication","dsmiv.opcrit","icd10.opcrit","panss.date"))
str(phenodat)

#PanssScore vs BMI 
title = "Relationship of Panss and BMI"
ggplot(data = filter(phenodat_melt,Phenotype!="control"), 
      aes(x = BMI, y=value, colour = Gender, fill = Gender)) +
      geom_point()+
      facet_wrap(~ variable,scales="free")+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))

#PanssScore vs Age vs Gender
title = "Relationship of Panss and Age"
ggplot(data = filter(phenodat_melt,Phenotype!="control"), 
      aes(x = Age, y=value, colour = Gender, fill = Gender)) +
      geom_point()+
      facet_wrap(~ variable,scales="free")+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))


#PanssScore vs Tobacco
title = "Relationship of Panss and Tobacco"
ggplot(data = filter(phenodat_melt,Phenotype!="control"), 
      aes(x = Tobacco, y=value, colour = Gender, fill = Gender)) +
      geom_boxplot(alpha = 0)+
      geom_jitter(alpha = 0.3)+
      #facet_wrap(~ variable,scales="free")+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))

#Panss Negative vs Panss Positive
title = "Relationship of PanssNegative and PanssPositive"
ggplot(data = filter(phenodat,Phenotype!="control"), 
      aes(x = PanssPositive, y=PanssNegative, colour = Gender, fill = Gender)) +
      geom_point()+
      ggtitle(title)
ggsave(paste(P0_figs_dir,title,".png",sep=""))


```
