==> R/p0_Split_Data_GAP_FEP_data_07_02_2017.Rmd <==
---
title: "Feature Selection"
author: "DJL"
date: "02/10/2016"
output:
  html_document:
    toc: yes
    toc_float: yes
---

#Split Data approach:

#Libraries
```{r load_libs, tidy=TRUE}
rm(list = ls())
dev.off()
library(lumi)
library(randomForest)
library(lubridate)
library(doMC)
library(caret)
library(plyr)
library(dplyr)
library(tableone)

```


##Set directories:
```{r directory}

P0_output_dir <-"./P0_Characterise/output/"
P0_figs_dir <-"./P0_Characterise/figs/"

```



#Functions
```{r define functions, tidy=TRUE}

table_fun<- function(mldata,split,prs_col=prs_col){
  listVars <- c("Gender","Age", "Ethnicity","BMI","Tobacco",names(mldata)[prs_col])
  catVars <- c("Ethnicity","Tobacco")
  table1 <- CreateTableOne(vars = listVars, data = mldata, factorVars = catVars,strata=c("Phenotype"),includeNA = T)
  table1print<-print(table1)
  write.table(table1print, file=paste(P0_output_dir,split,"_Table_1_Demographics.tsv",sep=""),row.names=T,quote=FALSE,sep = "\t")


  listVars2 <- c("Medication","dsmiv.opcrit","icd10.opcrit","PanssScore","PanssPositive","PanssNegative","PanssPsycho")
  catVars2 <- c("Medication","dsmiv.opcrit","icd10.opcrit")
  table2 <- CreateTableOne(vars = listVars2, data = filter(mldata,Phenotype=="FEP"), factorVars = catVars2,includeNA = T)
  table2print<-print(table2)
  write.table(table2print, file=paste(P0_output_dir,split,"_Table_2_Clinicalinformation.tsv",sep=""),row.names=T,quote=FALSE,sep = "\t")
}
```



#Load Data
```{r Load data}

load(paste(P0_output_dir,"GAP_FEP_eset_linear_adj_Data.RData",sep=""))
#load(paste(P0_output_dir,"GAP_FEP_eset_linear_full_adj_Data.RData",sep=""))


```

#Define Cores and Seed
```{r Define Cores and Seed}
#Allow 8 Cores
registerDoMC(cores = 8) 
#Set Seed
seed = 7
set.seed(seed)

```


##Split data
```{r}

###########################################
## Eset linear_adj split 
#This is the data that had cell types regressed out.
GX_lumi_object<-eset_linear_adj


# remove duplicate probes and use gene symbols
gene<- t(exprs(GX_lumi_object))
probenames<-GX_lumi_object@featureData@data
colnames(gene) <- probenames$TargetID
gene_expression_matrix<-t(gene)
exprs_data<-gene
dim(exprs_data)

# pheno data
names(pData(GX_lumi_object))
pheno_data<-pData(GX_lumi_object)[,c(1,6,9,11:19,74,21:24,64:73,75:84)]
names(pheno_data)


set.seed(seed)

names(pheno_data)
#Add phenotype to Gene expression. 
GX_DF<-droplevels(merge(pheno_data,exprs_data, by.x="sampleID",by.y="row.names"))
row.names(GX_DF)<-GX_DF[,1]
GX_DF[1:10,1:50]
GX_DF_adj<-GX_DF

training_index <- createDataPartition(GX_DF$Phenotype, p = 0.70, list = F)

training_df <- GX_DF[training_index,]
testing_df <- GX_DF[-training_index,]

#Make tables
split="testing_df"
table_fun(testing_df,split,prs_col=(18:26))
split="training_df"
table_fun(training_df,split,prs_col=(18:26))


################################################
## Eset linear_full_adj split 
#This is the data that had cell types and ethnicity gender sex regressed out
GX_lumi_object<-NULL
GX_lumi_object<-eset_linear_full_adj


# remove duplicate probes and use gene symbols
gene<- t(exprs(GX_lumi_object))
probenames<-GX_lumi_object@featureData@data
colnames(gene) <- probenames$TargetID
gene_expression_matrix<-t(gene)
exprs_data<-gene
dim(exprs_data)

# pheno data
names(pData(GX_lumi_object))
pheno_data<-pData(GX_lumi_object)[,c(1,6,9,11:19,74,21:24,64:73)]
names(pheno_data)


set.seed(seed)

#Add phenotype to Gene expression. 
GX_DF<-droplevels(merge(pheno_data,exprs_data, by.x="sampleID",by.y="row.names"))
row.names(GX_DF)<-GX_DF[,1]
GX_DF[1:10,1:10]
GX_DF_full_adj<-GX_DF

training_index <- createDataPartition(GX_DF$Phenotype, p = 0.70, list = F)

training_df_full <- GX_DF[training_index,]
testing_df_full <- GX_DF[-training_index,]

#Make tables
split="testing_df_full"
table_fun(testing_df_full,split,prs_col=(18:27))
split="training_df_full"
table_fun(training_df_full,split,prs_col=(18:27))

###########################################################################
#Save results
save(training_df, testing_df,training_df_full, testing_df_full,file=paste(P0_output_dir,"Train_and_Test_data.Rdata",sep=""), compress = T)
save(GX_DF_full_adj,file=paste(P0_output_dir,"GX_DF_full_adj_data.Rdata",sep=""), compress = T)
save(GX_DF_adj,file=paste(P0_output_dir,"GX_DF_adj_data.Rdata",sep=""), compress = T)





```
