==> R/p1_4_GAP_Variables_test_07_02_2017.Rmd <==
---
title: "Variable_imp_lee_vs_gap"
author: "DJL"
date: "06/06/2016"
output: html_document
---
#Todo
```{r todo list, tidy=TRUE}
1: impute missing data
2: 
```



#Libraries
```{r load_libs, tidy=TRUE}
library(tableone)
library(dplyr)
```


##Set directories:
```{r directory}
P1_output_dir <-"./P1_Hypothesis_Free/output/"
P1_figs_dir <-"./P1_Hypothesis_Free/figs/"
```


## Check most important variables for GAP
```{r}

#use model to predict training data
predict_train_fit<-predict.train(RF_final, newdata = testing_df)
confusionMatrix(predict_train_fit, testing_df$Phenotype)

#check variables
names(testing_df)[1:26]

#create object to be used from now on. 
full_data<-testing_df[,1:26]


#add rownames to predictions
tpredictions<-data.frame(predict_train_fit)
rownames(tpredictions)<-full_data$sampleID


#create test data with pheno predictions. 
model_test_pheno<-merge(tpredictions,full_data,by.x="row.names",by.y="sampleID")
## Make confusion matrix column
model_test_pheno<-model_test_pheno %>% mutate(ConfusionMatrix=ifelse(Phenotype == "FEP" & predict_train_fit=="control","FEP_misclassed",
                                            ifelse(Phenotype == "FEP" & predict_train_fit=="FEP","FEP_true",
                                            ifelse(Phenotype != "FEP" & predict_train_fit=="FEP","Control_misclassified","Control_true"))))
# Center and Scale PRS
model_test_pheno[19:27]<-scale(model_test_pheno[19:27])


```


## Create Table of variables
```{r}


#Create Table of demographics with 4 predicted groups. 
listVars <- c("Gender","Age", "Ethnicity","BMI","Tobacco",names(model_test_pheno)[19:27])
catVars <- c("Ethnicity","Tobacco")
table1_full <- CreateTableOne(vars = listVars, data = model_test_pheno, factorVars = catVars,strata=c("ConfusionMatrix"),includeNA = T)
table1print<-print(table1_full)
write.table(table1print, file=paste(P1_output_dir,"p1_4_RF_model_Table_CM_1_Demographics.tsv",sep=""),row.names=T,quote=FALSE,sep = "\t")

#Create Table of demographics with 4 predicted groups. 
listVars2 <- c("Medication","dsmiv.opcrit","icd10.opcrit","PanssScore","PanssPositive","PanssNegative","PanssPsycho")
catVars2 <- c("Medication","dsmiv.opcrit","icd10.opcrit")
table2_full <- CreateTableOne(vars = listVars2, data = filter(model_test_pheno,Phenotype=="FEP"), factorVars = catVars2,strata=c("ConfusionMatrix"),includeNA = T)
table2print<-print(table2_full)
write.table(table2print, file=paste(P1_output_dir,"p1_4_RF_model_Table_CM_1_meds_Demographics.tsv",sep=""),row.names=T,quote=FALSE,sep = "\t")


```



## Create Table of variables
```{r}

dim(model_test_pheno)
FEP_data<-filter(model_test_pheno,Phenotype=="FEP")


SUMO3<-testing_df$SUMO3

FEP_data<-testing_df
FEP_data<-merge(tpredictions,FEP_data,by.x="row.names",by.y="sampleID")




FEP_data<-filter(FEP_data,Phenotype=="FEP")
FEP_data<-filter(FEP_data,Phenotype=="FEP")
FEP_data<-FEP_data[complete.cases(FEP_data$Pol_1_GAP_all_strict_excl_WTCCC2),]
FEP_data[19:27]<-scale(FEP_data[19:27])
corPRS_genes<-cor(FEP_data[,-c(1:26)], method = c("pearson"))

?sort
sort(corPRS_genes[1,],decreasing = T)["SUMO3"]

```
