==> R/p11_GAP_Bootstrap_Enrichment_07_02_2017.Rmd <==
---
title: "Machine Learning"
author: "DJL"
date: "02/10/2016"
output:
  html_document:
    toc: yes
    toc_float: yes
---

#Overview

**Aim**  
Create Machine learning models using GAP data. 


#Libraries
```{r load_libs, tidy=TRUE}
rm(list = ls())
dev.off()
library(plyr)
library(dplyr)
library(lubridate)
library(reshape)
library(tableone)
library(stargazer)
library(VennDiagram)
library(WGCNA)
library(flashClust)
library(data.table)
```

#Functions
```{r define functions, tidy=TRUE}
#venn Diagram function
plotVennDia <- function(a, ...) {
    grid.newpage()
    if (length(a) == 1) {
        out <- draw.single.venn(likes(a), ...)
    }
    if (length(a) == 2) {
        out <- draw.pairwise.venn(likes(a[1]), likes(a[2]), likes(a[1:2]), ...)
    }
    if (length(a) == 3) {
        out <- draw.triple.venn(likes(a[1]), likes(a[2]), likes(a[3]), likes(a[1:2]), 
            likes(a[2:3]), likes(a[c(1, 3)]), likes(a), ...)
    }
    if (length(a) == 4) {
        out <- draw.quad.venn(likes(a[1]), likes(a[2]), likes(a[3]), likes(a[4]), 
            likes(a[1:2]), likes(a[c(1, 3)]), likes(a[c(1, 4)]), likes(a[2:3]), 
            likes(a[c(2, 4)]), likes(a[3:4]), likes(a[1:3]), likes(a[c(1, 2, 
                4)]), likes(a[c(1, 3, 4)]), likes(a[2:4]), likes(a), ...)
    }
    if (!exists("out")) 
        out <- "Oops"
    return(out)
}

#Venn Diagram helper fun
likes <- function(animals) {
    ppl <- allpredictorsUnique
    for (i in 1:length(animals)) {
        ppl <- subset(ppl, ppl[animals[i]] == T)
    }
    nrow(ppl)
}

#likes(c("Gx","Gx_Scz","Gx_PRS"))


```


#Directories
```{r Define directories}
top_dir<-getwd()
data_dir <-"./P0_Characterise/output/"
data_dir0 <-"./data/"
output_dir <-"./P11_glmnet_bootstrap/output/"
figs_dir <-"./P11_glmnet_bootstrap/figs/"
Enrichment_dir <- "./P11_glmnet_bootstrap/Enrichment/"



##### Set up #####
#project name defined
project_name = "glmnet_enrichment"
project_id = "p11_"

```

#  Load predictors
```{r Predictor lists}

#Load data
glmnet_all_predictors<-read.csv(paste(output_dir,"p11_glmnet_boot_glmnet_all_mod_predictors.csv",sep=""))

#change value 1 to PRS_0.1 needed for true false df for venn diagram. 
glmnet_all_predictors[2]<-replace(glmnet_all_predictors[2], glmnet_all_predictors[2]==1, "PRS_0.1")

#vector of model ids
modelID<-c("1_Gx","2_Gx_Scz","3_PRS","4_Demographics","5_PRS_Demo","6_Gx_PRS","7_Gx_Demo","8_Gx_Demo_PRS")

```



# Venn Diagram 
```{r Enrichment}
#Make venn diagram of overlap. For Gx, Gx_scz, PRS, PRS_Gx?


#drop factors
allpredictors <- data.frame(lapply(glmnet_all_predictors, as.character), stringsAsFactors=F)

#Get unique predictors
allpredictorsUnique<-unique(allpredictors[2])




#loop to add false true for each
for (i in 1:length(modelID)){
  temp<-filter(glmnet_all_predictors, Model == modelID[i])[2]
  allpredictorsUnique[,modelID[i]]<-allpredictorsUnique[,1] %in% temp[,1]
}

#change rownames
rownames(allpredictorsUnique)<-allpredictorsUnique[,1]

#define new colnames
new_colnames<-c("Variables","Gx","Gx_Scz","PRS","Demographics","PRS_Demo","Gx_PRS","Gx_Demo","Gx_Demo_PRS")

#colnames chage
colnames(allpredictorsUnique)<-new_colnames


#Venn Diagram 3 way
jpeg(paste(figs_dir,project_id,project_name,"VennDiagram_3.jpg",sep=""),width= 800, height = 800, pointsize = 20)
plotVennDia(c("Gx", "Gx_Scz","Gx_PRS"), category = c("Gx", "Gx_Scz","Gx_PRS"), lty = "blank", fill = c("skyblue", "pink1","mediumorchid"))
dev.off()

#Venn Diagram 4 way
jpeg(paste(figs_dir,project_id,project_name,"VennDiagram_4.jpg",sep=""),width= 800, height = 800, pointsize = 20)
plotVennDia(c("Gx", "Gx_Scz","Gx_PRS","Gx_Demo"), category = c("Gx", "Gx_Scz","Gx_PRS","Gx_Demo"), lty = "blank", fill = c("skyblue", "pink1","mediumorchid", "orange"))
dev.off()


```




## User List Enrichement for Gx probes Scz later
```{r gene_list}
#load background data
all_probes_file<-"BackgroundEnrichment.tsv"
all_probes<-read.csv(paste(data_dir0,all_probes_file,sep=""),sep="\t",header=TRUE)

#create Gx subset
Gx_probes<-allpredictorsUnique[allpredictorsUnique$Gx == T,]


#Add Background column
all_probes$GxBackground<-all_probes$TargetID%in%Gx_probes$Variables
all_probes$GxBackground<-replace(all_probes$GxBackground,all_probes$GxBackground=="FALSE","BACKGROUND")
all_probes$GxBackground<-replace(all_probes$GxBackground,all_probes$GxBackground=="TRUE","Gx")


setwd(top_dir)
setwd(Enrichment_dir)
enrichments = userListEnrichment(all_probes$TargetID,all_probes$GxBackground,fnIn=c("GAP_reduced_GO_Biological_Process_2015.csv","GAP_reduced_GO_Cellular_Component_2015.csv","GAP_reduced_GO_Molecular_Function_2015.csv","GAP_reduced_Kegg2016.csv","GeneSetsMental_Pirooznia_2016final2.csv","GAP_reduced_Blood_WGCNA.csv","GAP_reduced_Brain_WGCNA.csv"),catNmIn=c("GO_BP","GO_CC","GO_MF","KEGG_2016","Pirooznia","Blood","Brain"),minGenesInCategory = 3)
setwd(top_dir)
enpv<-enrichments$pValue
enrichment_probes<-unlist(lapply(enrichments$ovGenes,paste, collapse =";"))
enpv$Genes<-enrichment_probes
enpv<-enpv[order(enpv$InputCategories,enpv$Pvalues),]
enpvDT<-data.table(enpv)
enpvDT<-enpvDT[Pvalues < 0.05 & NumOverlap > 1,.SD[],by=InputCategories]

write.csv(as.data.frame(enpvDT),file=paste(Enrichment_dir,project_id,project_name,"_Gx_enrichment_results.csv",sep=""),row.names=F)



```

## User List Enrichement for Gx_Scz
```{r gene_list}
#load background data
all_probes_file<-"BackgroundEnrichment.tsv"
all_probes<-read.csv(paste(data_dir0,all_probes_file,sep=""),sep="\t",header=TRUE)

#create Gx subset
Gx_probes<-allpredictorsUnique[allpredictorsUnique$Gx_Scz == T,]



#Add Background column
all_probes$GxBackground<-all_probes$TargetID%in%Gx_probes$Variables
all_probes$GxBackground<-replace(all_probes$GxBackground,all_probes$GxBackground=="FALSE","BACKGROUND")
all_probes$GxBackground<-replace(all_probes$GxBackground,all_probes$GxBackground=="TRUE","Gx")


setwd(top_dir)
setwd(Enrichment_dir)
enrichments = userListEnrichment(all_probes$TargetID,all_probes$GxBackground,fnIn=c("GAP_reduced_GO_Biological_Process_2015.csv","GAP_reduced_GO_Cellular_Component_2015.csv","GAP_reduced_GO_Molecular_Function_2015.csv","GAP_reduced_Kegg2016.csv","GeneSetsMental_Pirooznia_2016final2.csv","GAP_reduced_Blood_WGCNA.csv","GAP_reduced_Brain_WGCNA.csv"),catNmIn=c("GO_BP","GO_CC","GO_MF","KEGG_2016","Pirooznia","Blood","Brain"),minGenesInCategory = 3)
setwd(top_dir)
enpv<-enrichments$pValue
enrichment_probes<-unlist(lapply(enrichments$ovGenes,paste, collapse =";"))
enpv$Genes<-enrichment_probes
enpv<-enpv[order(enpv$InputCategories,enpv$Pvalues),]
enpvDT<-data.table(enpv)
enpvDT<-enpvDT[Pvalues < 0.05 & NumOverlap > 1,.SD[],by=InputCategories]

write.csv(as.data.frame(enpvDT),file=paste(Enrichment_dir,project_id,project_name,"_Gx_scz_enrichment_results.csv",sep=""),row.names=F)



```
